@@ -1,4 +1,4 @@
-class CreatePerson < ActiveRecord::Migration
+class CreatePerson < ActiveRecord::Migration[4.2]
   def change
     create_table :people do |t|
       t.string :name, null: false
