@@ -1,4 +1,4 @@
-class CreateNotes < ActiveRecord::Migration
+class CreateNotes < ActiveRecord::Migration[4.2]
   def change
     create_table :notes do |t|
       t.references :person, index: true, null: false
