@@ -1,4 +1,4 @@
-class AddFileToAttachment < ActiveRecord::Migration
+class AddFileToAttachment < ActiveRecord::Migration[4.2]
   def change
     add_column :attachments, :file_id, :string
     add_column :attachments, :file_filename, :string
