@@ -1,4 +1,4 @@
-class RemovePriorityToPerson < ActiveRecord::Migration
+class RemovePriorityToPerson < ActiveRecord::Migration[4.2]
   def change
     remove_column :people, :priority
   end
