@@ -1,4 +1,4 @@
-class AddRoleToUser < ActiveRecord::Migration
+class AddRoleToUser < ActiveRecord::Migration[4.2]
   def change
     add_column :users, :role, :string, default: 'admin', null: false
   end
