@@ -1,4 +1,4 @@
-class RemoveAssessments < ActiveRecord::Migration
+class RemoveAssessments < ActiveRecord::Migration[4.2]
   def change
     drop_table :assessments
   end
