@@ -1,4 +1,4 @@
-class AddStatusToPerson < ActiveRecord::Migration
+class AddStatusToPerson < ActiveRecord::Migration[4.2]
   def change
     add_column :people, :status, :string
   end
