@@ -1,4 +1,4 @@
-class AddUpdatedByIdToDayoffs < ActiveRecord::Migration
+class AddUpdatedByIdToDayoffs < ActiveRecord::Migration[4.2]
   def change
     add_column :dayoffs, :updated_by_id, :integer
   end
