@@ -1,4 +1,4 @@
-class DefaultStatus < ActiveRecord::Migration
+class DefaultStatus < ActiveRecord::Migration[4.2]
   def change
     change_column :people, :status, :string, default: 'n/a', null: false
 
