@@ -1,4 +1,4 @@
-class AddSourceToPerson < ActiveRecord::Migration
+class AddSourceToPerson < ActiveRecord::Migration[4.2]
   def change
     add_column :people, :source, :string
   end
