@@ -1,4 +1,4 @@
-class AddVacationOverrideToPerson < ActiveRecord::Migration
+class AddVacationOverrideToPerson < ActiveRecord::Migration[4.2]
   def change
     add_column :people, :vacation_override, :integer
   end
