@@ -1,4 +1,4 @@
-class AddStatusToVacancy < ActiveRecord::Migration
+class AddStatusToVacancy < ActiveRecord::Migration[4.2]
   def change
     add_column :vacancies, :status, :string, default: 'open'
   end
