@@ -3,6 +3,6 @@
 //= require bootstrap
 //= require bootstrap-datepicker
 //= require select2
-//= require jquery-ui/autocomplete
+//= require jquery-ui
 //= require autocomplete-rails
 //= require_tree .
